299 - Validações parte 1 - Controle de fluxos

+++ app/Http/Controllers/MarcaController.php
@@ -14,7 +14,6 @@ class MarcaController extends Controller
         $this->marca = $marca;
     }
 
-
     /**
      * Display a listing of the resource.
      */
@@ -22,7 +21,6 @@ class MarcaController extends Controller
     {
         // $marcas = Marca::all();
         $marcas = $this->marca::all();
-
         return $marcas;
     }
 
@@ -53,6 +51,11 @@ class MarcaController extends Controller
     {
         $marca = $this->marca->find($id);
 
+        if ($marca === null) {
+            return ['erro' => 'Recurso pesquisado não existe'];
+        }
+
+
         return $marca;
     }
 
@@ -70,15 +73,13 @@ class MarcaController extends Controller
     public function update(Request $request, $id)
     {
 
-        // print_r($request->all()); //os dados atualizados
-        // echo '<hr>';
-        // print_r($marca->getAttributes()); // os dados antigos
-
-        // $marca->update($request->all());
-
         $marca = $this->marca->find($id);
-        $marca->update($request->all());
 
+        if ($marca === null) {
+            return ['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe'];
+        }
+
+        $marca->update($request->all());
         return $marca;
     }
 
@@ -87,9 +88,13 @@ class MarcaController extends Controller
      */
     public function destroy($id)
     {
-        // $marca->delete();
 
         $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return ['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe'];
+        }
+
         $marca->delete();
 
         return ['msg' => 'A marca foi removida com sucesso!'];
